+++ packages/plugin-tailwind/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Minor Changes
 
+- Fully supports DTCG 2025.10 and resolvers.
+
+- ⚠️ Breaking change: Minor config change from `modeVariations` → `variations`. The behavior is backwards-compatible with 0.x but requires a config update.
+
 - [#530](https://github.com/terrazzoapp/terrazzo/pull/530) [`370ed7b`](https://github.com/terrazzoapp/terrazzo/commit/370ed7b0f578a64824124145d7f4936536b37bb3) Thanks [@drwpow](https://github.com/drwpow)! - ⚠️ Breaking change: lint on plugins no longer runs on individual files, rather, the full set once merged.
 
   If your lint plugin is not using the `src` context value, no changes are needed. If it is, you’ll need to instead read from the `sources` array, and look up sources with a token’s `source.loc` filename manually. This change was because lint rules now run on all files in one pass, essentially.